Add --seed_start argument and tidy up seed handling
The --seed_start argument allows you to say "Please use the following
seed for the first iteration of each test. If you run more iterations
than one, increment the seed accordingly".

The idea is that some top-level program (for example Ibex's core_ibex
UVM test code) can pick a seed, possibly randomly, and then run many
tests using the same seeds for test generation (here) and test
execution (passed to the simulator that runs them).

We could implement that by generating a YAML file each time and
passing it with --seed_yaml, but that seems unnecessarily complicated.

While touching the seed handling code, I've also added checks for
various silly combinations of arguments that used to be possible.

  - If you specify --seed and --iterations 10, you'll now get an
    error. This is probably sensible because you're asking for 10
    identical tests, which is unlikely to be what you want.

  - If you specify --seed and don't specify --iterations, the
    iteration count is now forced to 1 rather than being read from the
    configuration file. Same reason as above.

The patch also pulls seed handling into a wrapper class (SeedGen),
which avoids having to pass yet another argument to do_simulate.

It also folds config file parsing into the end of parse_args and gets
rid of the call to var(..) at the end of load_config: we used to
convert from the args namespace to a dictionary called cfg, and then
promptly converted it back into a namespace. This seems a little
silly, so the patch gets rid of it.

Note: Before and after this patch, test seeds depend on the batch
size. This is because we don't re-seed between tests in a batch. For
example, these two commands will generate different programs:

    ./run.py --iterations=100 --batch-size 5 --seed_start 123
    ./run.py --iterations=100 --batch-size 10 --seed_start 123

It's not *too* bad, because the "test_id" which we use as a key in the
stored list of seeds is actually "test_name + batch index". But you'll
only be able to reproduce a generated test with a high iteration
number if you keep the batch size equal to the last time you ran.

This is easy enough to fix by passing an initial seed as a
plusarg (rather than using -svseed / -sv_seed arguments) and then
reseeding with, say, <seed> + <iteration> between each generated test.
I'll leave this for a later patch.