Add different methods to initialize the vregs
Three different methods have been implemented to initialize the vector registers.
We have mantained the current one that uses the instructions vmv and vslide1up and added
the method that was implemented before that used only the vmv instruction and implemented
a new method that uses the vle instruction.

This feature has been discussed in the issue #603.

This work has been done by the verification team in the EPI project at BSC-CNS.