Adds new PMP directed sequence.
This patch adds in a new directed sequence of instructions placed
directly after the PMP setup.
This directed sequence will write randomized values to every `pmpaddr`
and `pmpcfg` CSR in order to test write accessibility (and spec
compliance). Correctness checking for this sequence is determined by
post-comparison with the ISS log, as per usual.