fix jumps to `test_done` and `init_[m/s/u]_mode`
From #709, it appears there is an issue jumping to `init_machine_mode`
when generating a large (200k instructions) program.
This is because we currently use a standard `j` instruction to do the
jump which only has a 20-bit address range, so this range is exceeded in
a program of this size.
Similar issues arise when jumping to `test_done` from various exception
handlers.

The solution is as follows:

1) Move the `init_[m/s/u]_mode` section to the top of the program to
prevent us from having to jump there in the first place - this does not
cause any side effects.
2) Change the jump mechanisms to `test_done` from `j test_done` to
preloading the address of `test_done` into a register and using a `jalr`
instruction to jump to this address - this allows us to cover a full
XLEN-bit address space when jumping around the program.