debug_rom_gen: Fix return address issue
The debug ROM is assembled with the sub-routines directly below the main
code section. The program is thus able to run straight through into the
sub-routine without there ever being a jalr into that sub-routine. That
means the return address for the sub-routine is undefined, and the
program jumps to a bogus address when the sub-routine returns.

This fix forces the main debug program to end (by jumping to debug_end)
and not just run on into the sub-routines, similarly to how the main
(non-debug) program works.

Signed-off-by: Tom Roberts <tomroberts@lowrisc.org>