Allow for underscores and capital letters in ISA for ISS
Newer versions of Spike need the required bitmanip sub-extensions to be
specified separately in the ISA string as _Zba_Zbb_Zbc_Zbc_Xbitmanip
(where Xbitmanip comprises all sub-extensions not part of v.1.0.0 but
remaining in draft v.0.9.4) instead of just b. Previously, run.py would
print an error message when parsing such an ISA string but still call
Spike with the correct arguments which is confusing.

With this commit, run.py is modified to accept such ISA strings and not
print the error message.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>