Allow for WFI in User Mode
This commit allows the case because if mstatus.tw is low, WFI behaviour
will not change regarding the mode. If it's set to one, it will act like
an illegal instruction. That is also not a problem because we already
generate illegal instruction handler inside gen_all_trap_handler method.

Signed-off-by: Canberk Topal <ctopal@lowrisc.org>