[scripts] Improve WARL support in gen_csr_test
This adds a new kind of write field. It uses a fragment of python from
the CSR YAML to legalize a written value. This enables better support of
WARL fields (gen_csr_test can now be told how to predict a particular
RISC-V implementation will deal with illegal values that are written).

Signed-off-by: Greg Chadwick <gac@lowrisc.org>