[pmp] Add knob to suppress PMP setup code
By default PMP setup is included in every test which sets up a full set
of PMP regions. A test may wish to have full access to all of memory to
ensure there are no PMP failures.  This knob suppresses the general PMP
setup generation and instead generates a setup with a single region
giving full access to all of memory.