Add plusarg to enable ECALL insn in main randomized body
This allows us to get coverage over this instruction, but requires you to
rethink how the test_done mechanism works. Currently the ECALL handler is
responsible for cleanup and jumping to the end-of-test loop, so if you don't
change this mechanism the test will end upon reaching the first ECALL
instruction.
e.g. override the gen_test_done method in combination with enabling
this option.