Reserve one extra word when pushing GPRs to kernel stack
This is slightly cleaner as it means nothing will be stored at the starting
value of the KSP. If a processor comes out of reset and immediately jumps to
debug mode, the first code in the debug_rom is to push the GPRs to the kernel
stack. The stack pointer resets to 0x0, and has not yet been initialized.
This means the GPRs will be pushed to 0xffffff80 through 0x00000000 (inclusive),
which can for example cross a PMP boundary.
It's a slightly obscure problem but for the cost of one extra word in the stack
this is an easier fix.