Fix `update_src_regs` for ZBB
`update_src_regs` is used by coverage collection to extract source
operands. It doesn't work for most of the I format ZBB instructions as
they don't use the immediate as an operand but rather as extra bits to
determine operation. This introduces a special case `update_src_regs` to
the ZBB instruction group to handle this.