Remove support for stale vector amo extension
Support for vector amo instructions has been removed from the standard vector extension. The encoding of the instructions also collides with the scalar subword atomic endcoding and no new encoding is currently available. Once there is further progress on Zvamo, the instructions can be added back and updated to the newest version